+++ packages/client/src/lib/drizzle/schema.ts
@@ -1,3 +1,4 @@
+export * from "./schema/checkpoints";
 export * from "./schema/directorUpdates";
 export * from "./schema/llamaInferences";
 export * from "./schema/simulations";